@@ -1,32 +1,42 @@
 name: moleculis
 description: Moleculis flutter application
 
-
 version: 0.3.0+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  http: ^0.12.2
-  shared_preferences: ^0.5.8
-  flutter_bloc: ^6.0.1
-  equatable: ^1.2.3
-  easy_localization: ^2.3.2
-  google_fonts: ^1.1.0
-  package_info: ^0.4.1
-  auto_size_text: ^2.1.0
-  settings_ui: ^0.3.0
-  image_picker: ^0.6.7+4
-  cached_network_image: ^2.2.0+1
+  cupertino_icons: ^1.0.3
+  http: ^0.13.3
+  shared_preferences: ^2.0.6
+  flutter_bloc: ^7.0.1
+  equatable: ^2.0.2
+  easy_localization: ^3.0.0
+  google_fonts: ^2.1.0
+  package_info: ^2.0.2
+  auto_size_text: ^3.0.0-nullsafety.0
+  settings_ui: ^1.0.0
+  image_picker: ^0.8.0+3
+  cached_network_image: ^3.0.0
+  flexible_scrollbar: ^0.1.2
+  get_it: ^7.1.3
+  firebase_core: ^1.3.0
+  cloud_firestore: ^2.2.2
+  firebase_storage: ^8.1.3
+  freezed_annotation: ^0.14.2
+  firebase_database: ^7.1.1
+  firebase_messaging: ^10.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.7.5
+  flutter_launcher_icons: ^0.9.0
+  build_runner: ^2.0.4
+  freezed: ^0.14.2
+  json_serializable: ^4.1.3
 
 flutter_icons:
   ios: true
