@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:moleculis/blocs/authentication/authentication_bloc.dart';
+import 'package:moleculis/blocs/auth/auth_bloc.dart';
 import 'package:moleculis/blocs/events/events_bloc.dart';
 import 'package:moleculis/blocs/events/events_event.dart';
 import 'package:moleculis/blocs/events/events_state.dart';
@@ -16,26 +16,29 @@ import 'package:moleculis/screens/event_details/widgets/users_list.dart';
 import 'package:moleculis/widgets/gradient_button.dart';
 import 'package:moleculis/widgets/info_item.dart';
 import 'package:moleculis/widgets/input.dart';
-import 'package:moleculis/widgets/loading_widget.dart';
+import 'package:moleculis/widgets/loading_wrapper.dart';
 import 'package:moleculis/widgets/select_date_time.dart';
 import 'package:moleculis/widgets/simple_button.dart';
 import 'package:moleculis/widgets/toolbar.dart';
 
 class CreateEditEventScreen extends StatefulWidget {
-  final int eventId;
   final EventsBloc eventsBloc;
+  final int? eventId;
 
-  const CreateEditEventScreen({Key key, this.eventId, this.eventsBloc})
-      : super(key: key);
+  const CreateEditEventScreen({
+    Key? key,
+    required this.eventsBloc,
+    this.eventId,
+  }) : super(key: key);
 
   @override
   _CreateEditEventScreenState createState() => _CreateEditEventScreenState();
 }
 
 class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
-  EventsBloc eventsBloc;
-  Event event;
-  User currentUser;
+  late final EventsBloc eventsBloc;
+  Event? event;
+  User? currentUser;
 
   final GlobalKey<FormState> formKey = GlobalKey();
 
@@ -48,188 +51,182 @@ class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
   final TextEditingController locationController = TextEditingController();
   final FocusNode locationFocusNode = FocusNode();
 
-  DateTime date = DateTime.now();
+  DateTime? date = DateTime.now();
 
-  File pickedImage;
+  File? pickedImage;
 
   final List<UserSmall> newUsers = [];
 
   @override
   void initState() {
     eventsBloc = widget.eventsBloc;
-    currentUser =
-        BlocProvider.of<AuthenticationBloc>(context).state.currentUser;
+    currentUser = BlocProvider.of<AuthBloc>(context).state.currentUser;
     if (widget.eventId != null) {
-      event = eventsBloc.getEventById(widget.eventId);
-      titleController.text = event.title;
-      descriptionController.text = event.description;
-      locationController.text = event.location;
+      event = eventsBloc.getEventById(widget.eventId!);
+      titleController.text = event!.title;
+      descriptionController.text = event!.description!;
+      locationController.text = event!.location!;
 
-      date = event.date;
+      date = event!.date;
     }
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return BlocListener<EventsBloc, EventsState>(
-      cubit: eventsBloc,
-      listener: (BuildContext context, EventsState state) {
-        if (state is EventsSuccess) {
-          Navigator.pop(context);
-        }
-      },
-      child: Scaffold(
-        body: BlocBuilder<EventsBloc, EventsState>(
-            cubit: eventsBloc,
-            builder: (BuildContext context, EventsState eventsState) {
-              return SafeArea(
-                child: Form(
-                  key: formKey,
-                  child: Stack(
-                    children: <Widget>[
-                      SingleChildScrollView(
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 20,
-                            vertical: 20,
-                          ),
-                          child: Column(
-                            children: <Widget>[
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Toolbar(
-                                  title: (event != null
-                                      ? 'edit_event'
-                                      : 'create_event')
-                                      .tr(),
-                                  backButton: true,
-                                  onImagePicked: (File image) {
-                                    pickedImage = image;
-                                  },
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Input(
-                                  controller: titleController,
-                                  focusNode: titleFocusNode,
-                                  nextFocusNode: descriptionFocusNode,
-                                  textInputAction: TextInputAction.next,
-                                  onFieldSubmitted: (String value) {},
-                                  validator: (String value) {
-                                    value = value.trim();
-                                    if (value.isEmpty) {
-                                      return 'title_empty'.tr();
-                                    }
-                                    if (value.length < 5) {
-                                      return 'title_short'.tr();
-                                    }
-                                    return null;
-                                  },
-                                  title: 'title'.tr(),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Input(
-                                  controller: descriptionController,
-                                  focusNode: descriptionFocusNode,
-                                  nextFocusNode: locationFocusNode,
-                                  textInputAction: TextInputAction.next,
-                                  onFieldSubmitted: (String value) {},
-                                  title: 'description'.tr(),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Input(
-                                  controller: locationController,
-                                  focusNode: locationFocusNode,
-                                  textInputAction: TextInputAction.next,
-                                  onFieldSubmitted: (String value) {},
-                                  validator: (String value) {
-                                    value = value.trim();
-                                    if (value.isEmpty) {
-                                      return 'location_empty'.tr();
-                                    }
-                                    if (value.length < 5) {
-                                      return 'location_short'.tr();
-                                    }
-                                    return null;
-                                  },
-                                  title: 'location'.tr(),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: SelectDateTime(
-                                  onTap: () => selectDate(context),
-                                  selectedDate: date,
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: SelectDateTime(
-                                  onTap: () => selectTime(context),
-                                  title: '${'time'.tr()}:',
-                                  isTime: true,
-                                  selectedDate: date,
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: InfoItem(
-                                  title: 'participants'.tr(),
-                                  contentWidget: Column(
-                                    mainAxisSize: MainAxisSize.min,
-                                    children: <Widget>[
-                                      UsersList(
-                                        editing: true,
-                                        onStateChane: () {
-                                          setState(() {});
-                                        },
-                                        users: event?.users ?? newUsers,
-                                      ),
-                                      Container(
-                                        width: double.infinity,
-                                        child: SimpleButton(
-                                          text: '+',
-                                          onPressed: onPickUsersTapped,
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: GradientButton(
-                                  onPressed: save,
-                                  text: 'save'.tr(),
-                                ),
-                              ),
-                            ],
+    return Scaffold(
+      body: BlocConsumer<EventsBloc, EventsState>(
+        bloc: eventsBloc,
+        listener: (BuildContext context, EventsState state) {
+          if (state is EventsSuccess) {
+            Navigator.pop(context);
+          }
+        },
+        builder: (BuildContext context, EventsState eventsState) {
+          return LoadingWrapper(
+            isLoading: eventsState.isLoading,
+            child: SafeArea(
+              child: Form(
+                key: formKey,
+                child: SingleChildScrollView(
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 20,
+                      vertical: 20,
+                    ),
+                    child: Column(
+                      children: <Widget>[
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Toolbar(
+                            title:
+                                (event != null ? 'edit_event' : 'create_event')
+                                    .tr(),
+                            backButton: true,
+                            onImagePicked: (File image) {
+                              pickedImage = image;
+                            },
                           ),
                         ),
-                      ),
-                      if (eventsState.isLoading) LoadingWidget(),
-                    ],
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Input(
+                            controller: titleController,
+                            focusNode: titleFocusNode,
+                            nextFocusNode: descriptionFocusNode,
+                            textInputAction: TextInputAction.next,
+                            onFieldSubmitted: (String value) {},
+                            validator: (String? value) {
+                              value = value!.trim();
+                              if (value.isEmpty) {
+                                return 'title_empty'.tr();
+                              }
+                              if (value.length < 5) {
+                                return 'title_short'.tr();
+                              }
+                              return null;
+                            },
+                            title: 'title'.tr(),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Input(
+                            controller: descriptionController,
+                            focusNode: descriptionFocusNode,
+                            nextFocusNode: locationFocusNode,
+                            textInputAction: TextInputAction.next,
+                            onFieldSubmitted: (String value) {},
+                            title: 'description'.tr(),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Input(
+                            controller: locationController,
+                            focusNode: locationFocusNode,
+                            textInputAction: TextInputAction.next,
+                            onFieldSubmitted: (String value) {},
+                            validator: (String? value) {
+                              value = value!.trim();
+                              if (value.isEmpty) {
+                                return 'location_empty'.tr();
+                              }
+                              if (value.length < 5) {
+                                return 'location_short'.tr();
+                              }
+                              return null;
+                            },
+                            title: 'location'.tr(),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: SelectDateTime(
+                            onTap: () => selectDate(context),
+                            selectedDate: date,
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: SelectDateTime(
+                            onTap: () => selectTime(context),
+                            title: '${'time'.tr()}:',
+                            isTime: true,
+                            selectedDate: date,
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: InfoItem(
+                            title: 'participants'.tr(),
+                            contentWidget: Column(
+                              mainAxisSize: MainAxisSize.min,
+                              children: <Widget>[
+                                UsersList(
+                                  editing: true,
+                                  onStateChane: () {
+                                    setState(() {});
+                                  },
+                                  users: event?.users ?? newUsers,
+                                ),
+                                Container(
+                                  width: double.infinity,
+                                  child: SimpleButton(
+                                    text: '+',
+                                    onPressed: onPickUsersTapped,
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: GradientButton(
+                            onPressed: save,
+                            text: 'save'.tr(),
+                          ),
+                        ),
+                      ],
+                    ),
                   ),
                 ),
-              );
-            }),
+              ),
+            ),
+          );
+        },
       ),
     );
   }
 
   Future<void> onPickUsersTapped() async {
-    final User resultUser = await showSearch(
+    final User? resultUser = await showSearch<User?>(
       context: context,
       delegate: UserSearch(
         excludeUsername: event != null
-            ? event.users.map((e) => e.username).toList()
-            : [currentUser.username],
+            ? event!.users.map((e) => e.username).toList()
+            : [currentUser!.username],
       ),
     );
     if (resultUser != null) {
@@ -240,9 +237,9 @@ class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
   }
 
   Future<void> selectDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
         context: context,
-        initialDate: date,
+        initialDate: date!,
         firstDate: DateTime(2015, 8),
         lastDate: DateTime(2050));
     if (picked != null && picked != date)
@@ -252,16 +249,16 @@ class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
   }
 
   Future<void> selectTime(BuildContext context) async {
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay? picked = await showTimePicker(
       context: context,
-      initialTime: TimeOfDay.fromDateTime(date),
+      initialTime: TimeOfDay.fromDateTime(date!),
     );
-    if (picked != null && picked != TimeOfDay.fromDateTime(date))
+    if (picked != null && picked != TimeOfDay.fromDateTime(date!))
       setState(() {
         date = DateTime(
-          date.year,
-          date.month,
-          date.day,
+          date!.year,
+          date!.month,
+          date!.day,
           picked.hour,
           picked.minute,
         );
@@ -269,7 +266,7 @@ class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
   }
 
   void save() {
-    if (formKey.currentState.validate()) {
+    if (formKey.currentState!.validate()) {
       final request = CreateUpdateEventRequest(
         title: titleController.text,
         description: descriptionController.text,
@@ -286,11 +283,11 @@ class _CreateEditEventScreenState extends State<CreateEditEventScreen> {
       } else {
         eventsBloc.add(
           UpdateEvent(
-            widget.eventId,
+            widget.eventId!,
             request,
-            event.users
+            event!.users
               ..where(
-                    (element) => element.username != currentUser.username,
+                (element) => element.username != currentUser!.username,
               ),
           ),
         );
