@@ -3,11 +3,11 @@ import 'dart:io';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:moleculis/blocs/authentication/authentication_bloc.dart';
+import 'package:moleculis/blocs/auth/auth_bloc.dart';
 import 'package:moleculis/blocs/groups/groups_bloc.dart';
 import 'package:moleculis/blocs/groups/groups_event.dart';
 import 'package:moleculis/blocs/groups/groups_state.dart';
-import 'package:moleculis/models/group/group.dart';
+import 'package:moleculis/models/group.dart';
 import 'package:moleculis/models/requests/create_update_group_request.dart';
 import 'package:moleculis/models/user/user.dart';
 import 'package:moleculis/models/user/user_small.dart';
@@ -16,18 +16,18 @@ import 'package:moleculis/screens/event_details/widgets/users_list.dart';
 import 'package:moleculis/widgets/gradient_button.dart';
 import 'package:moleculis/widgets/info_item.dart';
 import 'package:moleculis/widgets/input.dart';
-import 'package:moleculis/widgets/loading_widget.dart';
+import 'package:moleculis/widgets/loading_wrapper.dart';
 import 'package:moleculis/widgets/simple_button.dart';
 import 'package:moleculis/widgets/toolbar.dart';
 
 class CreateEditGroupScreen extends StatefulWidget {
-  final int groupId;
+  final int? groupId;
   final GroupsBloc groupsBloc;
 
   const CreateEditGroupScreen({
-    Key key,
+    Key? key,
     this.groupId,
-    @required this.groupsBloc,
+    required this.groupsBloc,
   }) : super(key: key);
 
   @override
@@ -35,9 +35,10 @@ class CreateEditGroupScreen extends StatefulWidget {
 }
 
 class _CreateEditGroupScreenState extends State<CreateEditGroupScreen> {
-  Group group;
-  User currentUser;
-  GroupsBloc groupsBloc;
+  late final User currentUser;
+  late final GroupsBloc groupsBloc;
+
+  Group? group;
 
   final GlobalKey<FormState> formKey = GlobalKey();
 
@@ -50,171 +51,165 @@ class _CreateEditGroupScreenState extends State<CreateEditGroupScreen> {
   final List<UserSmall> newUsers = [];
   final List<UserSmall> newAdmins = [];
 
-  File pickedImage;
+  File? pickedImage;
 
   @override
   void initState() {
     groupsBloc = widget.groupsBloc;
-    currentUser =
-        BlocProvider.of<AuthenticationBloc>(context).state.currentUser;
+    currentUser = BlocProvider.of<AuthBloc>(context).state.currentUser!;
     if (widget.groupId != null) {
-      group = groupsBloc.getGroupById(widget.groupId);
-      titleController.text = group.title;
-      descriptionController.text = group.description;
+      group = groupsBloc.getGroupById(widget.groupId!);
+      titleController.text = group!.title;
+      descriptionController.text = group!.description!;
     }
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return BlocListener<GroupsBloc, GroupsState>(
-      cubit: groupsBloc,
-      listener: (BuildContext context, GroupsState state) {
-        if (state is GroupsSuccess) {
-          Navigator.pop(context);
-        }
-      },
-      child: Scaffold(
-        body: BlocBuilder<GroupsBloc, GroupsState>(
-            cubit: groupsBloc,
-            builder: (BuildContext context, GroupsState eventsState) {
-              return SafeArea(
-                child: Form(
-                  key: formKey,
-                  child: Stack(
-                    children: <Widget>[
-                      SingleChildScrollView(
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 20,
-                            vertical: 20,
-                          ),
-                          child: Column(
-                            children: <Widget>[
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Toolbar(
-                                  title: (group != null
-                                          ? 'edit_group'
-                                          : 'create_group')
-                                      .tr(),
-                                  backButton: true,
-                                  onImagePicked: (File image) {
-                                    pickedImage = image;
-                                  },
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Input(
-                                  controller: titleController,
-                                  focusNode: titleFocusNode,
-                                  nextFocusNode: descriptionFocusNode,
-                                  textInputAction: TextInputAction.next,
-                                  onFieldSubmitted: (String value) {},
-                                  validator: (String value) {
-                                    value = value.trim();
-                                    if (value.isEmpty) {
-                                      return 'title_empty'.tr();
-                                    }
-                                    if (value.length < 5) {
-                                      return 'title_short'.tr();
-                                    }
-                                    return null;
-                                  },
-                                  title: 'title'.tr(),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: Input(
-                                  controller: descriptionController,
-                                  focusNode: descriptionFocusNode,
-                                  textInputAction: TextInputAction.next,
-                                  onFieldSubmitted: (String value) {},
-                                  title: 'description'.tr(),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: InfoItem(
-                                  title: 'participants'.tr(),
-                                  contentWidget: Column(
-                                    mainAxisSize: MainAxisSize.min,
-                                    children: <Widget>[
-                                      UsersList(
-                                        editing: true,
-                                        onStateChane: () {
-                                          setState(() {});
-                                        },
-                                        users: group?.users ?? newUsers,
-                                      ),
-                                      Container(
-                                        width: double.infinity,
-                                        child: SimpleButton(
-                                          text: '+',
-                                          onPressed: onPickUsersTapped,
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: InfoItem(
-                                  title: 'admins'.tr(),
-                                  contentWidget: Column(
-                                    mainAxisSize: MainAxisSize.min,
-                                    children: <Widget>[
-                                      UsersList(
-                                        editing: true,
-                                        showCurrentUser: true,
-                                        onStateChane: () {
-                                          setState(() {});
-                                        },
-                                        users: group?.admins ?? newAdmins,
-                                      ),
-                                      Container(
-                                        width: double.infinity,
-                                        child: SimpleButton(
-                                          text: '+',
-                                          onPressed: () =>
-                                              onPickUsersTapped(isAdmins: true),
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.only(bottom: 20),
-                                child: GradientButton(
-                                  onPressed: save,
-                                  text: 'save'.tr(),
-                                ),
-                              ),
-                            ],
+    return Scaffold(
+      body: BlocConsumer<GroupsBloc, GroupsState>(
+        bloc: groupsBloc,
+        listener: (BuildContext context, GroupsState state) {
+          if (state is GroupsSuccess) {
+            Navigator.pop(context);
+          }
+        },
+        builder: (BuildContext context, GroupsState eventsState) {
+          return LoadingWrapper(
+            isLoading: eventsState.isLoading,
+            child: SafeArea(
+              child: Form(
+                key: formKey,
+                child: SingleChildScrollView(
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 20,
+                      vertical: 20,
+                    ),
+                    child: Column(
+                      children: <Widget>[
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Toolbar(
+                            title:
+                                (group != null ? 'edit_group' : 'create_group')
+                                    .tr(),
+                            backButton: true,
+                            onImagePicked: (File image) {
+                              pickedImage = image;
+                            },
                           ),
                         ),
-                      ),
-                      if (eventsState.isLoading) LoadingWidget(),
-                    ],
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Input(
+                            controller: titleController,
+                            focusNode: titleFocusNode,
+                            nextFocusNode: descriptionFocusNode,
+                            textInputAction: TextInputAction.next,
+                            onFieldSubmitted: (String value) {},
+                            validator: (String? value) {
+                              value = value!.trim();
+                              if (value.isEmpty) {
+                                return 'title_empty'.tr();
+                              }
+                              if (value.length < 5) {
+                                return 'title_short'.tr();
+                              }
+                              return null;
+                            },
+                            title: 'title'.tr(),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: Input(
+                            controller: descriptionController,
+                            focusNode: descriptionFocusNode,
+                            textInputAction: TextInputAction.next,
+                            onFieldSubmitted: (String value) {},
+                            title: 'description'.tr(),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: InfoItem(
+                            title: 'participants'.tr(),
+                            contentWidget: Column(
+                              mainAxisSize: MainAxisSize.min,
+                              children: <Widget>[
+                                UsersList(
+                                  editing: true,
+                                  onStateChane: () {
+                                    setState(() {});
+                                  },
+                                  users: group?.users ?? newUsers,
+                                ),
+                                Container(
+                                  width: double.infinity,
+                                  child: SimpleButton(
+                                    text: '+',
+                                    onPressed: onPickUsersTapped,
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: InfoItem(
+                            title: 'admins'.tr(),
+                            contentWidget: Column(
+                              mainAxisSize: MainAxisSize.min,
+                              children: <Widget>[
+                                UsersList(
+                                  editing: true,
+                                  showCurrentUser: true,
+                                  onStateChane: () {
+                                    setState(() {});
+                                  },
+                                  users: group?.admins ?? newAdmins,
+                                ),
+                                Container(
+                                  width: double.infinity,
+                                  child: SimpleButton(
+                                    text: '+',
+                                    onPressed: () =>
+                                        onPickUsersTapped(isAdmins: true),
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 20),
+                          child: GradientButton(
+                            onPressed: save,
+                            text: 'save'.tr(),
+                          ),
+                        ),
+                      ],
+                    ),
                   ),
                 ),
-              );
-            }),
+              ),
+            ),
+          );
+        },
       ),
     );
   }
 
   Future<void> onPickUsersTapped({bool isAdmins = false}) async {
-    List<String> excludeUsers;
+    late List<String> excludeUsers;
     if (group != null) {
-      excludeUsers = (isAdmins ? group.admins : group.users)
+      excludeUsers = (isAdmins ? group!.admins : group!.users)
           .map((e) => e.username)
           .toList();
-      excludeUsers.addAll((isAdmins ? group.users : group.admins)
+      excludeUsers.addAll((isAdmins ? group!.users : group!.admins)
           .map((e) => e.username)
           .toList());
       if (!excludeUsers.contains(currentUser.username)) {
@@ -224,11 +219,9 @@ class _CreateEditGroupScreenState extends State<CreateEditGroupScreen> {
       excludeUsers = [currentUser.username];
     }
 
-    final User resultUser = await showSearch(
+    final User? resultUser = await showSearch<User?>(
       context: context,
-      delegate: UserSearch(
-        excludeUsername: excludeUsers,
-      ),
+      delegate: UserSearch(excludeUsername: excludeUsers),
     );
     if (resultUser != null) {
       setState(() {
@@ -242,7 +235,7 @@ class _CreateEditGroupScreenState extends State<CreateEditGroupScreen> {
   }
 
   void save() {
-    if (formKey.currentState.validate()) {
+    if (formKey.currentState!.validate()) {
       final CreateUpdateGroupRequest request = CreateUpdateGroupRequest(
         title: titleController.text,
         description: descriptionController.text,
@@ -251,8 +244,12 @@ class _CreateEditGroupScreenState extends State<CreateEditGroupScreen> {
       if (group == null) {
         groupsBloc.add(CreateGroupEvent(request, newUsers, newAdmins));
       } else {
-        groupsBloc.add(
-            UpdateGroupEvent(request, group.users, group.admins, group.id));
+        groupsBloc.add(UpdateGroupEvent(
+          request,
+          group!.users,
+          group!.admins,
+          group!.id,
+        ));
       }
     }
   }
