@@ -2,12 +2,12 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:moleculis/blocs/authentication/authentication_bloc.dart';
+import 'package:moleculis/blocs/auth/auth_bloc.dart';
 import 'package:moleculis/blocs/groups/groups_bloc.dart';
 import 'package:moleculis/blocs/groups/groups_state.dart';
-import 'package:moleculis/models/group/group.dart';
+import 'package:moleculis/models/group.dart';
 import 'package:moleculis/models/user/user.dart';
-import 'package:moleculis/models/user/user_small.dart';
+import 'package:moleculis/screens/chat/chat_screen.dart';
 import 'package:moleculis/screens/create_edit_group/create_edit_group_screen.dart';
 import 'package:moleculis/screens/event_details/widgets/users_list.dart';
 import 'package:moleculis/utils/navigation.dart';
@@ -20,9 +20,9 @@ class GroupDetailsScreen extends StatefulWidget {
   final GroupsBloc groupsBloc;
 
   const GroupDetailsScreen({
-    Key key,
-    @required this.groupId,
-    this.groupsBloc,
+    Key? key,
+    required this.groupId,
+    required this.groupsBloc,
   }) : super(key: key);
 
   @override
@@ -30,7 +30,7 @@ class GroupDetailsScreen extends StatefulWidget {
 }
 
 class _GroupDetailsScreenState extends State<GroupDetailsScreen> {
-  GroupsBloc groupsBloc;
+  late final GroupsBloc groupsBloc;
 
   @override
   void initState() {
@@ -40,11 +40,11 @@ class _GroupDetailsScreenState extends State<GroupDetailsScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final Group group = groupsBloc.getGroupById(widget.groupId);
+    final Group group = groupsBloc.getGroupById(widget.groupId)!;
     final User currentUser =
-        BlocProvider.of<AuthenticationBloc>(context).state.currentUser;
+        BlocProvider.of<AuthBloc>(context).state.currentUser!;
     bool isAdmin = false;
-    for (UserSmall user in group.admins) {
+    for (final user in group.admins) {
       if (user.username == currentUser.username) {
         isAdmin = true;
         break;
@@ -72,9 +72,9 @@ class _GroupDetailsScreenState extends State<GroupDetailsScreen> {
       ),
       body: SingleChildScrollView(
         child: BlocBuilder<GroupsBloc, GroupsState>(
-          cubit: groupsBloc,
+          bloc: groupsBloc,
           builder: (BuildContext context, GroupsState state) {
-            final Group group = groupsBloc.getGroupById(widget.groupId);
+            final Group group = groupsBloc.getGroupById(widget.groupId)!;
             return Padding(
               padding: const EdgeInsets.all(20),
               child: Column(
@@ -111,6 +111,16 @@ class _GroupDetailsScreenState extends State<GroupDetailsScreen> {
             );
           },
         ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigation.toScreen(
+            context: context,
+            screen: ChatScreen(group: group),
+          );
+        },
+        backgroundColor: Colors.blue,
+        child: Icon(Icons.chat, color: Colors.white),
       ),
     );
   }
