@@ -1,4 +1,4 @@
 class Validation {
-  static final RegExp emailRegExp = RegExp("""^[a-zA-Z0-9.!#\$%&’*+/=?^_`{|}~-]+
-  @[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*\$""");
+  static final RegExp emailRegExp = RegExp('''^[a-zA-Z0-9.!#\$%&’*+/=?^_`{|}~-]+
+  @[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*\$''');
 }
